+++ src/main/java/insane96mcp/mobspropertiesrandomness/data/json/condition/ConditionsRegistry.java
@@ -38,6 +38,7 @@ public class ConditionsRegistry {
         register("time_played", MPRTimePlayedCondition.class);
         register("dimension", MPRDimensionCondition.class);
         register("biome", MPRBiomeCondition.class);
+        register("block_on", MPRBlockOnCondition.class);
         register("structure", MPRStructureCondition.class);
         register("deepness", MPRDeepnessCondition.class);
         register("distance_from_spawn", MPRDistanceFromSpawnCondition.class);