+++ changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Upcoming
+* Added `block_on` condition
+  * Condition is fulfilled if the target is on the `blocks` (list of blocks or block tags) specified
 * Reworked Range object
   * `min` and `max` are now Modifiable Values objects, so you can apply modifiers to either of them
   * The range object still has `modifiers` that will apply to both `min` and `max`